backend :)

+++ backend/ecommerce-v2/src/main/java/orchowski/tomasz/ecommercev2/repositories/ItemRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item,Long> {
+public interface ItemRepository extends JpaRepository<Item,Integer> {
 
    Optional<Item> findById(int id);
 }
